Enable Browser Caching
Basset returns the default cache-control:no-cache header set by laravel, which is very undesirable for static assets. This basically breaks all notions for browser (or cdn) caching.
The fix is as simple as defining a few headers that tell browsers (and pull zone cdn's) to store the resource for a fixed time. This change does exactly that. The caching period is set to one day. Potentially this setting can be moved to basset's config.